Automatically Start/Stop CF Tunnels
With this commit tunnels are now automatically started/stopped when needed.

Instead of having to connect to a localhost address this mod will now create the tunnel with the hostname we're connecting to and automatically route the connection though the tunnel it just created it.

It can also automatically start using a tunnel based on DNS entries, or (as a fallback) the user can define hostnames in the `forced_tunnels.json` file. For this the user also no longer needs to define anything other than the hostname, as the mod will automatically do everything else.

DNS Spec:
We look for a TXT record on the same subdomain as the server we are connecting to.
If there is a TXT record with the value "cloudflared-use-tunnel" or "cloudflared-route=<route>" we know that we should use a CF tunnel to connect to the server.
If the server has the TXT record "cloudflared-route=<route>", it will use the route as the route for the tunnel.
If the server has the TXT record "cloudflared-use-tunnel", it will use the server address as the route for the tunnel.
If the server has neither of the TXT records, it will not use a tunnel. (unless it is in the forced tunnels list)

Note: Only one of the above records are needed.